fix: device groups bad serialization
